♻️ 아울렛을 포함한 레이아웃 컴포넌트를 Layout.tsx 파일에 모음 (#249)
